model/model_utils.py
import os
from transformers import VisionEncoderDecoderModel, ViTImageProcessor, AutoTokenizer
import torch
from PIL import Image
from huggingface_hub import snapshot_download
import io

# --- Заглушка для будущей мультимодальности ---
MODEL_NAME = "tuman/vit-rugpt2-image-captioning"
LOCAL_MODEL_DIR = os.path.join(os.path.dirname(__file__), "model_snapshot")

if not os.path.exists(LOCAL_MODEL_DIR):
    os.makedirs(LOCAL_MODEL_DIR, exist_ok=True)
    snapshot_download(
        repo_id=MODEL_NAME,
        local_dir=LOCAL_MODEL_DIR,
        local_dir_use_symlinks=False,
        force_download=True,
        resume_download=False
    )

model = VisionEncoderDecoderModel.from_pretrained(LOCAL_MODEL_DIR)
feature_extractor = ViTImageProcessor.from_pretrained(LOCAL_MODEL_DIR)
tokenizer = AutoTokenizer.from_pretrained(LOCAL_MODEL_DIR)
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
model.to(device)
max_length = 16
num_beams = 4
gen_kwargs = {"max_length": max_length, "num_beams": num_beams}

def predict_caption_single_image(image_bytes):
    i_image = Image.open(io.BytesIO(image_bytes))
    if i_image.mode != "RGB":
        i_image = i_image.convert(mode="RGB")
    pixel_values = feature_extractor(images=[i_image], return_tensors="pt").pixel_values.to(device)
    output_ids = model.generate(pixel_values, **gen_kwargs)
    pred = tokenizer.batch_decode(output_ids, skip_special_tokens=True)[0].strip()
    return pred

# --- Заглушки для мультимодальных функций ---
def predict_multimodal(*args, **kwargs):
    # TODO: реализовать мультимодальную генерацию
    return "[Мультимодальная генерация не реализована]"

def parse_features(features_text):
    # TODO: реализовать парсинг признаков для мультимодальных моделей
    return features_text
